drivers/location: Add location API
The Location API is designed to provide the essential location
data which can be produced from sources like SUPL. GNSS, WIFI-
location, cellular location, propriatary solutions etc.

It is primed for use with the upcomming multi interface device
model, and for implementation into the upcomming GNSS API
