--- subsys/location/CMakeLists.txt
@@ -1,4 +1,0 @@
-# Copyright (c) 2022, Bjarki Arge Andreasen
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_library_sources(location.c)
