--- subsys/location/Kconfig
@@ -1,24 +1,0 @@
-# Copyright (c) 2022, Bjarki Arge Andreasen
-# SPDX-License-Identifier: Apache-2.0
-
-config LOCATION
-	bool "Location subsystem"
-	help
-	  Include location subsystem
-
-if LOCATION
-
-config LOCATION_PROVIDERS_MAX
-	int "Max supported providers"
-	default 4
-	help
-	  Sets the maximum supported providers
-
-config LOCATION_EVENT_HANDLERS_MAX
-	int "Max supported location event handlers"
-	default 4
-	help
-	  Sets the maximum supported location event
-	  handlers
-
-endif
