--- tests/location/disabled/testcase.yaml
@@ -1,4 +1,0 @@
-tests:
-  location.api:
-    tags: location_api_only
-    platform_allow: native_posix
