--- tests/location/enabled/prj.conf
@@ -1,5 +1,0 @@
-CONFIG_LOCATION=y
-CONFIG_ZTEST=y
-CONFIG_USERSPACE=y
-CONFIG_LOCATION_PROVIDERS_MAX=2
-CONFIG_LOCATION_EVENT_HANDLERS_MAX=2
